// projects/todo/src/lib/state/todo-action-types.ts
export enum TodoActionTypes {
  GET_TODOS = '[Todo Resolver] Get Todos',
  GET_TODOS_SUCCESS = '[Todo Effect] Get todos success',
  GET_TODOS_FAILURE = '[Todo Effect] Get todos failure',

  ADD_TODO = '[Add Todo Component] Add Todo',
  ADD_TODO_SUCCESS = '[Add Todo Component] Add todo success',
  ADD_TODO_FAILURE = '[Add Todo Component] Add todo failure',

  EDIT_TODO = '[Edit Todo Component] Edit Todo',
  EDIT_TODO_SUCCESS = '[Edit Todo Component] Edit todo success',
  EDIT_TODO_FAILURE = '[Edit Todo Component] Edit todo failure',

  DELETE_TODO = '[Todos] Delete Todo',
  DELETE_TODO_SUCCESS = '[Todos] Delete todo success',
  DELETE_TODO_FAILURE = '[Todos] Delete todo failure',
}